feat(ui): shared save status indicator
