+++ docs/index.html
@@ -17,6 +17,7 @@
   <section id="loading"><div class="spinner"></div></section>
   <section id="appMessage" role="alert" aria-live="polite"></section>
   <main id="app">Cargando…</main>
+  <div id="save-status" class="fixed bottom-4 right-4 bg-gray-800 text-white text-sm py-2 px-4 rounded-lg shadow-lg opacity-0 transition-opacity duration-500"></div>
   <dialog id="dlgNuevoCliente" class="modal">
     <form method="dialog">
       <label for="nuevoClienteNombre">Nombre del cliente:</label>
